+++ Source/Scene/CameraSpindleController.js
@@ -337,7 +337,7 @@ define([
 
     CameraSpindleController.prototype._spin = function(movement) {
         if (this.mode === CameraSpindleControllerMode.AUTO) {
-            var point = this._camera.pickEllipsoid(this._ellipsoid, movement.startPosition);
+            var point = this._camera.pickEllipsoid(movement.startPosition, this._ellipsoid);
             if (typeof point !== 'undefined') {
                 this._pan(movement);
             } else {
@@ -375,8 +375,8 @@ define([
 
     CameraSpindleController.prototype._pan = function(movement) {
         var camera = this._camera;
-        var p0 = camera.pickEllipsoid(this._ellipsoid, movement.startPosition);
-        var p1 = camera.pickEllipsoid(this._ellipsoid, movement.endPosition);
+        var p0 = camera.pickEllipsoid(movement.startPosition, this._ellipsoid);
+        var p1 = camera.pickEllipsoid(movement.endPosition, this._ellipsoid);
 
         if (typeof p0 === 'undefined' || typeof p1 === 'undefined') {
             return;